(feat) Move tabBar logic to tabBar.js
